+++ caten_worship/models/__init__.py
@@ -3,6 +3,7 @@
 from .songs import songsDB
 from .dbx import get_dbx
 from .users import User
+from .users_profile import UserProfile
 from .mails import mail
 from .login_managers import login_manager