+++ caten_worship/models/users.py
@@ -1,5 +1,7 @@
 # models/users.py
 
+import datetime
+
 from flask import current_app
 
 from itsdangerous import TimedJSONWebSignatureSerializer as sign
@@ -20,9 +22,12 @@ class User(db.Model):
     email = db.Column(db.String(64), unique=True, nullable=False)
     password_hash = db.Column(db.String(64), nullable=False)
     displayname = db.Column(db.String(16), nullable=False)
+    register_time = db.Column(db.DateTime, nullable=False, default=datetime.datetime.today())
     is_authenticated = db.Column(db.Boolean, default=False)
     is_active = db.Column(db.Boolean, default=False)
     is_anonymous = db.Column(db.Boolean, default=False)
+    is_admin = db.Column(db.Boolean, default=False)
+    user_profile = db.relationship('UserProfile', backref='user', lazy=True)
 
     # 定義 password 為不可讀
     @property
@@ -42,7 +47,8 @@ class User(db.Model):
     # 產生帳號啟動碼
 
     def create_activate_token(self, expires_in=3600 * 24):
-        token_generator = sign(current_app.config['SECRET_KEY'], expires_in=expires_in)
+        token_generator = sign(
+            current_app.config['SECRET_KEY'], expires_in=expires_in)
         return token_generator.dumps({"user_id": self.id})
 
     # 取得用戶唯一識別碼